Fix/budget logic to goals

+++ src/main/java/eci/ieti/FinzenTransactionService/repository/IncomeRepository.java
@@ -3,12 +3,12 @@ package eci.ieti.FinzenTransactionService.repository;
 import eci.ieti.FinzenTransactionService.model.Income;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IncomeRepository extends JpaRepository<Income, Long> {
     List<Income> findByUserId(Long userId);
 
     @Query("SELECT COALESCE(SUM(i.amount), 0) FROM Income i WHERE i.userId = :userId")
-    Double sumByUserId(Long userId);
+    BigDecimal sumByUserId(Long userId);
 }
